Refactor into modular pipeline
